Add Terraform provider internal CI/CD
This work provides a Jenkins-based CI/CD workflow to both build and test
the CloudBolt Terraform Resource Provider. As part of that effort, it
uses a golang docker container to build the provider and downloads the
appropriate/compatible version of Terraform, arranges the resources as
needed on the filesystem, and uses the bundled order-blueprint example
to verify that the provider is functioning correctly.

[DEV-12937]


Former-commit-id: b0ef18ca4d3d7ef3f0fe9b70a7afb281aad25a09 [formerly b9d97346c4533fab6d2d629abba3cfdc61bc64ad] [formerly 55ccd0fb33c775c25331874dfbc369b0baafa4d3 [formerly 7bf1c78635f71b14bed1a2f8bc4367465244305e]]
Former-commit-id: c2dfb9be38afe6280c8a63ec81c7b454b0ae39b7
Former-commit-id: 636b826af4525e5d65404c9e9474077e80426adf