Add Terraform provider internal CI/CD
This work provides a Jenkins-based CI/CD workflow to both build and test
the CloudBolt Terraform Resource Provider. As part of that effort, it
uses a golang docker container to build the provider and downloads the
appropriate/compatible version of Terraform, arranges the resources as
needed on the filesystem, and uses the bundled order-blueprint example
to verify that the provider is functioning correctly.

[DEV-12937]